Feat/2bimestre

+++ src/controllers/LocacaoController.ts
@@ -6,9 +6,8 @@ class LocacaoController {
   async listar(request: Request, response: Response) {
     try {
       const catalogo = await locacaoSchema.find();
-      response
-        .status(200)
-        .json({ data: catalogo, error: false, msg: "Locações encontradas" });
+      response.status(200).json(catalogo);
+      // .json({ data: catalogo, error: false, msg: "Locações encontradas" });
     } catch (error) {
       response.status(400).json({
         data: error.message,
@@ -22,9 +21,8 @@ class LocacaoController {
     try {
       const { id } = request.params;
       const catalogo = await locacaoSchema.find({ _id: id });
-      response
-        .status(201)
-        .json({ data: catalogo, error: false, msg: "Locação encontrado" });
+      response.status(201).json(catalogo);
+      // .json({ data: catalogo, error: false, msg: "Locação encontrado" });
     } catch (error) {
       response.status(400).json({
         data: error.message,
@@ -47,11 +45,12 @@ class LocacaoController {
 
           const catalogo = await locacaoSchema.create(request.body);
 
-          response.status(201).json({
-            data: catalogo,
-            error: false,
-            msg: "Locação cadastrada com sucesso",
-          });
+          response.status(201).json(catalogo);
+          // .json({
+          //   data: catalogo,
+          //   error: false,
+          //   msg: "Locação cadastrada com sucesso",
+          // });
         } catch (error) {
           response.status(400).json({
             error: true,
@@ -75,18 +74,31 @@ class LocacaoController {
 
   async excluir(request: Request, response: Response) {
     try {
-      const { id } = request.params;
-      const catalogo = await locacaoSchema.findByIdAndDelete(id);
-      response.status(201).json({
-        data: catalogo,
-        error: false,
-        msg: "Locação excluída com sucesso",
-      });
+      const list = await filmeSchema.findById({ _id: request.body.filme });
+
+      if (list && list.status === "LOCADO") {
+        try {
+          await filmeSchema.updateOne(
+            { _id: request.body.filme },
+            { $set: { status: "DISPONÍVEL" } }
+          );
+
+          const { id } = request.params;
+          const catalogo = await locacaoSchema.findByIdAndDelete(id);
+          response.status(201).json(catalogo);
+        } catch (error) {
+          response.status(400).json({
+            data: error.message,
+            error: true,
+            msg: "Não foi possível excluir esta locação",
+          });
+        }
+      }
     } catch (error) {
       response.status(400).json({
         data: error.message,
         error: true,
-        msg: "Não foi possível excluir esta locação",
+        msg: "Não foi possível buscar o filme.",
       });
     }
   }
@@ -95,6 +107,24 @@ class LocacaoController {
     try {
       const { id } = request.params;
       const body = request.body;
+
+      let filmeAtual: any;
+
+      filmeAtual = await locacaoSchema.find({ _id: id });
+
+      await filmeSchema.updateOne(
+        { _id: filmeAtual[0].filme.join() },
+        { $set: { status: "DISPONÍVEL" } }
+      );
+
+      let filmeNovo: any;
+
+      filmeNovo = request.body.filme;
+
+      await filmeSchema.updateOne(
+        { _id: filmeNovo },
+        { $set: { status: "LOCADO" } }
+      );
 
       await locacaoSchema.findOneAndUpdate({ _id: id }, body, {
         returnOriginal: false,
